Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
tensorflow  # or tensorflow-cpu
Pillow        # PIL for image handling
numpy